requirements.txt
numpy>=1.21.5
scipy>=1.7.2
sklearn>=1.0.2
absl-py>=1.0.0
matplotlib>=3.5.0
typing_extensions>=4.1.1
